Change the way the network module deals with unknown and missing interfaces  (#1)
Changes to operstate polling and missing interfaces

New implementation will view all operstates that aren't "down" as being "up", enabling interfaces with "unknown" operstates to be recognised by the module. Interfaces which are absent from the system will continue to be polled at each update of the module, for tighter integration of hotswappable NICs.